+++ darts/tests/models/forecasting/test_foundation.py
@@ -8,7 +8,12 @@ import numpy as np
 import pytest
 
 from darts import TimeSeries, concatenate
-from darts.tests.conftest import TIREX_AVAILABLE, TORCH_AVAILABLE, tfm_kwargs
+from darts.tests.conftest import (
+    T0_AVAILABLE,
+    TIREX_AVAILABLE,
+    TORCH_AVAILABLE,
+    tfm_kwargs,
+)
 from darts.utils.likelihood_models import QuantileRegression
 from darts.utils.timeseries_generation import linear_timeseries
 
@@ -18,7 +23,13 @@ if not TORCH_AVAILABLE:
         allow_module_level=True,
     )
 
-from darts.models import Chronos2Model, PatchTSTFMModel, TimesFM2p5Model, TiRexModel
+from darts.models import (
+    Chronos2Model,
+    PatchTSTFMModel,
+    T0Model,
+    TimesFM2p5Model,
+    TiRexModel,
+)
 from darts.tests.models.forecasting.foundation_test_utils import (
     CHRONOS2_TINY_DIR,
     HF_HUB_DOWNLOAD_PATCH_TARGET,
@@ -27,6 +38,7 @@ from darts.tests.models.forecasting.foundation_test_utils import (
     TiRexStub,
     mock_hf_hub_download,
     timesfm2p5_tiny_context,
+    tiny_t0_dir,
 )
 
 
@@ -426,6 +438,17 @@ class TestFoundationModel:
                 )
             ]
             if TIREX_AVAILABLE
+            else []
+        )
+        + (
+            [
+                (
+                    T0Model,
+                    "*decoder*",
+                    {"local_dir": tiny_t0_dir()},
+                )
+            ]
+            if T0_AVAILABLE
             else []
         ),
     )